health bar wip

+++ Assets/Scripts/Health/Health.cs
@@ -130,7 +130,7 @@ namespace Health
             _rigidbody.MovePosition(_rigidbody.position + (knockBack*knockBackForce*Time.fixedDeltaTime));
         }
 
-        public float getCurrentHealth()
+        public int getCurrentHealth()
         {
             return currentHealth;
         }
